+++ flutter_inappwebview_macos/lib/src/web_authentication_session/web_authenticate_session.dart
@@ -57,6 +57,9 @@ class MacOSWebAuthenticationSession extends PlatformWebAuthenticationSession
   late final String? callbackURLScheme;
 
   @override
+  late final Map<String, String>? additionalHeaderFields;
+
+  @override
   late final WebAuthenticationSessionSettings? initialSettings;
 
   @override
@@ -70,12 +73,14 @@ class MacOSWebAuthenticationSession extends PlatformWebAuthenticationSession
   Future<MacOSWebAuthenticationSession> create({
     required WebUri url,
     String? callbackURLScheme,
+    Map<String, String>? additionalHeaderFields,
     WebAuthenticationSessionCompletionHandler onComplete,
     WebAuthenticationSessionSettings? initialSettings,
   }) async {
     var session = MacOSWebAuthenticationSession._create(
       url: url,
       callbackURLScheme: callbackURLScheme,
+      additionalHeaderFields: additionalHeaderFields,
       onComplete: onComplete,
       initialSettings: initialSettings,
     );
@@ -85,6 +90,10 @@ class MacOSWebAuthenticationSession extends PlatformWebAuthenticationSession
     args.putIfAbsent("id", () => session.id);
     args.putIfAbsent("url", () => session.url.toString());
     args.putIfAbsent("callbackURLScheme", () => session.callbackURLScheme);
+    args.putIfAbsent(
+      "additionalHeaderFields",
+      () => session.additionalHeaderFields,
+    );
     args.putIfAbsent("initialSettings", () => initialSettings?.toMap());
     await _staticChannel.invokeMethod('create', args);
     return session;
@@ -93,6 +102,7 @@ class MacOSWebAuthenticationSession extends PlatformWebAuthenticationSession
   MacOSWebAuthenticationSession._create({
     required this.url,
     this.callbackURLScheme,
+    this.additionalHeaderFields,
     this.onComplete,
     WebAuthenticationSessionSettings? initialSettings,
   }) : super.implementation(MacOSWebAuthenticationSessionCreationParams()) {