+++ flutter_inappwebview_macos/macos/flutter_inappwebview_macos/Sources/flutter_inappwebview_macos/WebAuthenticationSession/WebAuthenticationSession.swift
@@ -16,15 +16,17 @@ public class WebAuthenticationSession: NSObject, Disposable {
     var plugin: InAppWebViewFlutterPlugin?
     var url: URL
     var callbackURLScheme: String?
+    var additionalHeaderFields: [String: String]?
     var settings: WebAuthenticationSessionSettings
     var session: Any?
     var channelDelegate: WebAuthenticationSessionChannelDelegate?
     private var _canStart = true
     
-    public init(plugin: InAppWebViewFlutterPlugin, id: String, url: URL, callbackURLScheme: String?, settings: WebAuthenticationSessionSettings) {
+    public init(plugin: InAppWebViewFlutterPlugin, id: String, url: URL, callbackURLScheme: String?, additionalHeaderFields: [String: String]?, settings: WebAuthenticationSessionSettings) {
         self.id = id
         self.plugin = plugin
         self.url = url
+        self.additionalHeaderFields = additionalHeaderFields
         self.settings = settings
         super.init()
         self.callbackURLScheme = callbackURLScheme
@@ -41,6 +43,10 @@ public class WebAuthenticationSession: NSObject, Disposable {
     public func prepare() {
         if #available(macOS 10.15, *), let session = session as? ASWebAuthenticationSession {
             session.prefersEphemeralWebBrowserSession = settings.prefersEphemeralWebBrowserSession
+            if #available(macOS 14.4, *),
+               session.responds(to: Selector(("setAdditionalHeaderFields:"))) {
+                session.setValue(additionalHeaderFields, forKey: "additionalHeaderFields")
+            }
         }
     }